230411 : fix divide func

+++ main.c
@@ -3,7 +3,7 @@
 int add(int a, int b) {return a+b;}
 int minus(int a, int b ) {return a-b;}
 int multiply(int a, int b ) {return a*b;}
-int divide(int a, int b ) {return a/b;}
+float divide(int a, int b ) {return a/(float)b;}
 
 int main(void)
 {
